+++ bin/wordpos-cli.js
@@ -269,4 +269,3 @@ function sprint(results) {
     },'');
   }
 }
-